requirements.txt
flask
boto3
pytest
requests
python-dateutil
